+++ highlight.io/components/QuickstartContent/backend/rust/actix.tsx
@@ -59,7 +59,7 @@ async fn main() -> Result<(), Error> {
 		{
 			title: 'Verify your errors are being recorded.',
 			content:
-				"Now that you've set everything up, you can verify that the backend error handling works by throwing an error in a service. Visit the [highlight errors page](http://app.highlight.io/errors) and check that backend errors are coming in.",
+				"Now that you've set everything up, you can verify that the backend error handling works by throwing an error in a service. Visit the [highlight errors page](https://app.highlight.io/errors) and check that backend errors are coming in.",
 			code: [
 				{
 					text: `// ...