+++ sdk/highlight-node/src/index.ts
@@ -1,4 +1,4 @@
 export { Highlight } from './client'
 export * as Handlers from './handlers'
 export { H, HIGHLIGHT_REQUEST_HEADER } from './sdk'
-export type { HighlightContext, NodeOptions } from './types'
+export type { HighlightContext, NodeOptions, Metric } from './types'