+++ sdk/highlight-react/tsup.config.ts
@@ -2,11 +2,12 @@ import CssModulesPlugin from 'esbuild-css-modules-plugin'
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-	clean: true,
+	clean: false,
 	dts: true,
 	esbuildPlugins: [CssModulesPlugin()],
 	minify: true,
-	sourcemap: false,
-	splitting: false,
+	sourcemap: true,
+	splitting: true,
 	bundle: true,
+	treeshake: 'smallest',
 })